Fix Cygwin SSH-Agent startup issue

+++ .bashrc
@@ -8,8 +8,10 @@ if [[ -n "$( which ruby 2>/dev/null )" ]]; then
     done
 fi
 
-ps -W | awk '/ssh-agent/,NF=1' | xargs kill -f
+if [ "$( uname -o )" != "Cygwin" ]; then
+    ps -W | awk '/ssh-agent/,NF=1' | xargs kill -f
 
-{ eval `ssh-agent -s`; ssh-add ~/.ssh/*_rsa; } &>/dev/null
+    { eval `ssh-agent -s`; ssh-add ~/.ssh/*_rsa; } &>/dev/null
+fi;
 
 cd ~
